Change function return type
emu3_get_free_dir_content_blknum function returns a signed short that represents
a block number used for directories.

These values are used in the directories in the 7 blocks reserved for that purpose.
In this example, blocks `0x9` and `0xa` are being used.
7 blocks allow for 116 directory entries.

```
00001000  44 65 66 61 75 6c 74 20  46 6f 6c 64 65 72 20 20  |Default Folder  |
00001010  00 80 09 00 00 0a ff ff  ff ff ff ff ff ff ff ff  |................|
```

+++ emu3_fs.h
@@ -196,7 +196,7 @@ ssize_t emu3_listxattr(struct dentry *, char *, size_t);
 
 void emu3_free_dir_content_block(struct emu3_sb_info *, int);
 
-int emu3_get_free_dir_content_blknum(struct emu3_sb_info *);
+short emu3_get_free_dir_content_blknum(struct emu3_sb_info *);
 
 void emu3_set_fattrs(struct emu3_sb_info *, struct emu3_file_attrs *, loff_t);
